+++ R/write.R
@@ -380,7 +380,7 @@ write_data2 <- function(
   }
 
   if (!is.null(attr(data, "c_cm"))) {
-    warning("modifications with cm formulas are experimental. use at own risk")
+    # warning("modifications with cm formulas are experimental. use at own risk")
     c_cm <- attr(data, "c_cm")
   } else {
     c_cm <- ""
@@ -1377,7 +1377,8 @@ do_write_formula <- function(
   }
 
   if (array || cm) {
-    if (length(dfx$X) == 1) {
+    # only if single input
+    if (identical(dim(dfx), c(1L, 1L))) {
       attr(dfx, "f_ref") <- dims
     }
   }