+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Fixes
 
 * Restore `wb_add_image()` that was broken in 1.24 [#1598](https://github.com/JanMarvin/openxlsx2/pull/1598)
+* Formulas with `array = TRUE` now extend over the selected `dims` range. The `cm = TRUE` warning was removed.
 
 
 ***************************************************************************